requirements.txt
adbutils==1.0.8
airtest==1.2.6
apkutils2==1.0.0
attrs==22.1.0
bcrypt==4.0.0
BeautifulReport==0.1.3
cached-property==1.5.2
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
cigam==0.0.3
colorama==0.4.5
colored==1.4.3
comtypes==1.1.14
cryptography==37.0.4
cssselect==1.1.0
decorator==5.1.1
Deprecated==1.2.13
deprecation==2.1.0
docker==4.4.4
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
facebook-wda==1.4.6
filelock==3.8.0
gitdb2==2.0.6
GitPython==2.1.15
hrpc==1.0.9
idna==2.10
Jinja2==3.0.3
jsonschema==3.2.0
loguru==0.6.0
logzero==1.7.0
lxml==4.9.1
MarkupSafe==2.1.1
mss==6.1.0
nose==1.3.7
numpy==1.23.2
opencv-contrib-python==4.6.0.66
packaging==20.9
paramiko==2.11.0
parsel==1.6.0
Pillow==9.2.0
pocoui==1.0.87
progress==1.6
py==1.11.0
pyaml==19.12.0
pycparser==2.21
pyelftools==0.29
PyGithub==1.45
PyJWT==2.4.0
PyNaCl==1.5.0
pyparsing==3.0.9
pypiwin32==223
pyrsistent==0.18.1
python-dateutil==2.8.2
python-gitlab==1.15.0
pywin32==227
pywinauto==0.6.3
PyYAML==5.3.1
requests==2.24.0
retry==0.9.2
simple-tornado==0.2.2
simplejson==3.17.6
six==1.16.0
smmap==5.0.0
smmap2==3.0.1
svn==1.0.1
tabulate==0.8.10
texttable==1.6.4
tidevice==0.9.11
tornado==6.2
uiautomator2==2.16.19
unittestreport==1.5.1
urllib3==1.25.11
w3lib==2.0.1
websocket-client==0.48.0
whichcraft==0.6.1
win32-setctime==1.1.0
wrapt==1.14.1
xlwings==0.25.3
xmltodict==0.13.0
